Filter devices by serials, skip dialog if only one device is found
Teach the bcf tool to filter the list of discovered serial devices by
their serial strings. Devices whose serial strings do not start with
"bc-" will be silently ignored. A typical host will have a whole bunch
of serial ports and there is no need to present all of those to the
user.

If the filtered device list ends up with only one matching device, use
the device right away. There is no need to present the selection dialog
to the user in this case. This makes the bcf tool easier to use on
hosts where the name of the character device may change over time. The
bcf tool will always pick the only available BigClown device in this
case.