When a single bc-* device is found, select it automatically
This patch attempts to simplify what I believe might be the common
case where a developer has a single BigClown device connected to the
system. In that case, the bcf tool will automatically select the
device without asking the developer.

In all other cases, i.e., when either no device with the serial prefix
"bc-" is found, or when several such devices are found, the behavior
of the tool remains unchanged.