Do not unquote USB descriptor strings on retrieval
Some descriptors appear to contain "" (two quotes) if the value is
empty. This is most likely a result of a design problem in the tool that
wrote the values. Before, we would unquote such values. In the new
version we don't do that anymore to alert the user to the fact that the
value contains the quoting characters (it tends to confuse udev).