Make reset and bootloader sequences configurable
The bcf tool has built-in support for the reset and bootloader sequences
implemented by the Hardwario Tower Core Module. These sequences use the
DTR and RTS signals on the on-board FTDI chip to reboot the MCU or to
switch it into the bootloader mode.

I use a similar approach on the LoRa Module for firmware development
purposes. My reset and bootloader sequences are different since the
corresponding pins on the TypeABZ module are directly connected to DTR
and RTS signals (unlike in Core Module).

This patch makes it possible to use the bcf tool to flash firmware into
the Tower LoRa Module with an external FTDI converter and only minimal
HW modification in the module (boot and reset pins on the debug
connector directly connected to DTR and RTS). It adds support for
per-module reset and bootloader sequences selectable on the command line
with the new parameter -m. By default (if omitted), the parameter is set
to "Core". With -m LoRa, different reset and bootloader sequences
compatible with the LoRa Module will be used.